# swagger.json
{
  "consumes": [
    "application/json",
    "application/xml"
  ],
  "produces": [
    "application/json",
    "application/xml"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "title": "BrightBird API.",
    "contact": {
      "email": "venus@ipfsconmmunity.com"
    },
    "license": {
      "name": "MIT",
      "url": "http://opensource.org/licenses/MIT"
    },
    "version": "0.0.1"
  },
  "host": "192.168.200.103:12356",
  "basePath": "/api/v1",
  "paths": {
...
    "/job": {
      "post": {
        "description": "save job entity, create if not exist",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "job"
        ],
        "operationId": "saveJob",
        "parameters": [
          {
            "description": "job json",
            "name": "job",
            "in": "body",
            "required": true,
            "schema": {
              "description": "job json",
              "type": "object",
              "$ref": "#/definitions/job"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "myString",
            "schema": {
              "$ref": "#/definitions/myString"
            }
          },
          "503": {
            "description": "apiError",
            "schema": {
              "$ref": "#/definitions/apiError"
            }
          }
        }
      }
    },
    "/job/count": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "job"
        ],
        "summary": "Count all jobs by condition.",
        "operationId": "countJobRequest",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "description": "id of job",
            "name": "id",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "Name",
            "description": "name of job",
            "name": "name",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "listJobResp",
            "schema": {
              "$ref": "#/definitions/listJobResp"
            }
          },
          "503": {
            "description": "apiError",
            "schema": {
              "$ref": "#/definitions/apiError"
            }
          }
        }
      }
    },
    "/job/detail/{id}": {
      "get": {
        "description": "Get job detail by id",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "job"
        ],
        "operationId": "getJob",
        "parameters": [
          {
            "type": "string",
            "description": "job id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "jobDetailResp",
            "schema": {
              "$ref": "#/definitions/jobDetailResp"
            }
          },
          "503": {
            "description": "apiError",
            "schema": {
              "$ref": "#/definitions/apiError"
            }
          }
        }
      }
    },
    "/job/list": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "job"
        ],
        "summary": "Lists all jobs.",
        "operationId": "listJobs",
        "responses": {
          "200": {
            "description": "listJobResp",
            "schema": {
              "$ref": "#/definitions/listJobResp"
            }
          },
          "503": {
            "description": "apiError",
            "schema": {
              "$ref": "#/definitions/apiError"
            }
          }
        }
      }
    },
    "/job/next": {
      "get": {
        "description": "Get job schedule",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "job"
        ],
        "operationId": "jobNextNReq",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "id",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "N",
            "name": "n",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "int64Arr",
            "schema": {
              "$ref": "#/definitions/int64Arr"
            }
          },
          "503": {
            "description": "apiError",
            "schema": {
              "$ref": "#/definitions/apiError"
            }
          }
        }
      }
    },
    "/job/{id}": {
      "get": {
        "description": "Update job",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "job"
        ],
        "operationId": "updateJob",
        "parameters": [
          {
            "type": "string",
            "description": "job id",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "job update params",
            "name": "updateJobParams",
            "in": "body",
            "required": true,
            "schema": {
              "description": "job update params",
              "type": "object",
              "$ref": "#/definitions/updateJobRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "job",
            "schema": {
              "$ref": "#/definitions/job"
            }
          },
          "503": {
            "description": "apiError",
            "schema": {
              "$ref": "#/definitions/apiError"
            }
          }
        }
      },
      "delete": {
        "description": "Delete job by id",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "job"
        ],
        "operationId": "deleteJob",
        "parameters": [
          {
            "type": "string",
            "description": "id of  job",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "503": {
            "description": "apiError",
            "schema": {
              "$ref": "#/definitions/apiError"
            }
          }
        }
      }
    },
...
    "/run/{jobid}": {
      "post": {
        "description": "run job immediately",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "job"
        ],
        "operationId": "runJobImmediately",
        "parameters": [
          {
            "type": "string",
            "description": "job id",
            "name": "jobid",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "myString",
            "schema": {
              "$ref": "#/definitions/myString"
            }
          },
          "503": {
            "description": "apiError",
            "schema": {
              "$ref": "#/definitions/apiError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "PRMergedEventMatch": {
      "type": "object",
      "properties": {
        "basePattern": {
          "type": "string",
          "x-go-name": "BasePattern"
        },
        "repo": {
          "type": "string",
          "x-go-name": "Repo"
        },
        "sourcePattern": {
          "type": "string",
          "x-go-name": "SourcePattern"
        }
      },
      "x-go-package": "github.com/ipfs-force-community/brightbird/models"
    },
...
    "TagCreateEventMatch": {
      "type": "object",
      "properties": {
        "repo": {
          "type": "string",
          "x-go-name": "Repo"
        },
        "tagPattern": {
          "type": "string",
          "x-go-name": "TagPattern"
        }
      },
      "x-go-package": "github.com/ipfs-force-community/brightbird/models"
    },
...
    "apiError": {
      "description": "APIError",
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "x-go-name": "Message"
        }
      },
      "x-go-name": "APIError",
      "x-go-package": "github.com/ipfs-force-community/brightbird/models"
    },
...
    "int64Arr": {
      "description": "Int64Array",
      "type": "array",
      "items": {
        "type": "integer",
        "format": "int64"
      },
      "x-go-name": "Int64Array",
      "x-go-package": "github.com/ipfs-force-community/brightbird/models"
    },
    "job": {
      "description": "Job",
      "type": "object",
      "properties": {
        "createTime": {
          "description": "创建时间",
          "type": "string",
          "format": "int64",
          "x-go-name": "CreateTime"
        },
        "cronExpression": {
          "type": "string",
          "x-go-name": "CronExpression"
        },
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "execCount": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ExecCount"
        },
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "jobType": {
          "type": "string",
          "x-go-name": "JobType"
        },
        "modifiedTime": {
          "description": "最后修改时间",
          "type": "string",
          "format": "int64",
          "x-go-name": "ModifiedTime"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "prMergedEventMatchs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PRMergedEventMatch"
          },
          "x-go-name": "PRMergedEventMatchs"
        },
        "tagCreateEventMatchs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TagCreateEventMatch"
          },
          "x-go-name": "TagCreateEventMatchs"
        },
        "testFlowId": {
          "type": "string",
          "x-go-name": "TestFlowId"
        },
        "versions": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-name": "Versions"
        }
      },
      "x-go-name": "Job",
      "x-go-package": "github.com/ipfs-force-community/brightbird/models"
    },
    "jobDetailResp": {
      "description": "JobDetailResp",
      "type": "object",
      "properties": {
        "createTime": {
          "description": "创建时间",
          "type": "string",
          "format": "int64",
          "x-go-name": "CreateTime"
        },
        "cronExpression": {
          "type": "string",
          "x-go-name": "CronExpression"
        },
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "execCount": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ExecCount"
        },
        "groupName": {
          "type": "string",
          "x-go-name": "GroupName"
        },
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "jobType": {
          "type": "string",
          "x-go-name": "JobType"
        },
        "modifiedTime": {
          "description": "最后修改时间",
          "type": "string",
          "format": "int64",
          "x-go-name": "ModifiedTime"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "prMergedEventMatchs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PRMergedEventMatch"
          },
          "x-go-name": "PRMergedEventMatchs"
        },
        "tagCreateEventMatchs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TagCreateEventMatch"
          },
          "x-go-name": "TagCreateEventMatchs"
        },
        "testFlowId": {
          "type": "string",
          "x-go-name": "TestFlowId"
        },
        "testFlowName": {
          "type": "string",
          "x-go-name": "TestFlowName"
        },
        "versions": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-name": "Versions"
        }
      },
      "x-go-name": "JobDetailResp",
      "x-go-package": "github.com/ipfs-force-community/brightbird/models"
    },
...
    "listJobResp": {
      "description": "ListJobResp",
      "type": "array",
      "items": {
        "$ref": "#/definitions/job"
      },
      "x-go-name": "ListJobResp",
      "x-go-package": "github.com/ipfs-force-community/brightbird/models"
    },
...
    "myString": {
      "description": "MyString repretation string",
      "type": "string",
      "x-go-name": "MyString",
      "x-go-package": "github.com/ipfs-force-community/brightbird/models"
    },
...
    "updateJobRequest": {
      "description": "UpdateJobRequest",
      "type": "object",
      "properties": {
        "cronExpression": {
          "type": "string",
          "x-go-name": "CronExpression"
        },
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "testFlowId": {
          "type": "string",
          "x-go-name": "TestFlowId"
        },
        "versions": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-name": "Versions"
        }
      },
      "x-go-name": "UpdateJobRequest",
      "x-go-package": "github.com/ipfs-force-community/brightbird/models"
    }
  },
...
}